requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
faiss-cpu
langchain-community
requests
numpy
